Add MultiLineInput in useInputAttrs ; improve InsertScriptEditor
